updated package/imports for gradle (#26)
Jay and I worked together to create the command line for the front end. We allowed a user to use an integer to specify if they want to use a file path or a list of integers. We also allowed them to add delimeter of their choosing if they want to.